\name{showLAGOexample}
\alias{showLAGOexample}
\title{Show Mechanism Behind Toy Example}
\description{This function shows how the toy data from 
\code{\link{generateLAGOexample}} are generated.}
\usage{
showLAGOexample()
}
\author{Wanhua Su and Mu Zhu, University of Waterloo, Canada.}
\keyword{misc}

